Replaced Python dir with Python3 dir
