+++ xbmc/interfaces/python/AddonPythonInvoker.cpp
@@ -49,14 +49,14 @@
 
 #define RUNSCRIPT_SETUPTOOLS_HACK \
   "" \
-  "import imp,sys\n" \
+  "import types,sys\n" \
   "pkg_resources_code = \\\n" \
   "\"\"\"\n" \
   "def resource_filename(__name__,__path__):\n" \
   "  return __path__\n" \
   "\"\"\"\n" \
-  "pkg_resources = imp.new_module('pkg_resources')\n" \
-  "exec pkg_resources_code in pkg_resources.__dict__\n" \
+  "pkg_resources = types.ModuleType('pkg_resources')\n" \
+  "exec(pkg_resources_code, pkg_resources.__dict__)\n" \
   "sys.modules['pkg_resources'] = pkg_resources\n" \
   ""
 
@@ -77,11 +77,11 @@
 #endif
 
 namespace PythonBindings {
-  void initModule_xbmcgui(void);
-  void initModule_xbmc(void);
-  void initModule_xbmcplugin(void);
-  void initModule_xbmcaddon(void);
-  void initModule_xbmcvfs(void);
+  PyObject *PyInit_Module_xbmcgui(void);
+  PyObject *PyInit_Module_xbmc(void);
+  PyObject *PyInit_Module_xbmcplugin(void);
+  PyObject *PyInit_Module_xbmcaddon(void);
+  PyObject *PyInit_Module_xbmcvfs(void);
 }
 
 using namespace PythonBindings;
@@ -94,18 +94,24 @@ typedef struct
 
 static PythonModule PythonModules[] =
   {
-    { "xbmcgui",    initModule_xbmcgui    },
-    { "xbmc",       initModule_xbmc       },
-    { "xbmcplugin", initModule_xbmcplugin },
-    { "xbmcaddon",  initModule_xbmcaddon  },
-    { "xbmcvfs",    initModule_xbmcvfs    }
+    { "xbmcgui",    PyInit_Module_xbmcgui    },
+    { "xbmc",       PyInit_Module_xbmc       },
+    { "xbmcplugin", PyInit_Module_xbmcplugin },
+    { "xbmcaddon",  PyInit_Module_xbmcaddon  },
+    { "xbmcvfs",    PyInit_Module_xbmcvfs    }
   };
 
 #define PythonModulesSize sizeof(PythonModules) / sizeof(PythonModule)
 
 CAddonPythonInvoker::CAddonPythonInvoker(ILanguageInvocationHandler *invocationHandler)
   : CPythonInvoker(invocationHandler)
-{ }
+{
+    PyImport_AppendInittab("xbmcgui",    PyInit_Module_xbmcgui);
+    PyImport_AppendInittab("xbmc",       PyInit_Module_xbmc);
+    PyImport_AppendInittab("xbmcplugin", PyInit_Module_xbmcplugin);
+    PyImport_AppendInittab("xbmcaddon",  PyInit_Module_xbmcaddon);
+    PyImport_AppendInittab("xbmcvfs",    PyInit_Module_xbmcvfs);
+}
 
 CAddonPythonInvoker::~CAddonPythonInvoker() = default;
