+++ xbmc/interfaces/python/PythonInvoker.h
@@ -27,6 +27,8 @@
 #include "threads/CriticalSection.h"
 #include "threads/Event.h"
 
+typedef struct _object PyObject;
+
 class CPythonInvoker : public ILanguageInvoker
 {
 public:
@@ -34,11 +36,10 @@ public:
   ~CPythonInvoker() override;
 
   bool Execute(const std::string &script, const std::vector<std::string> &arguments = std::vector<std::string>()) override;
-
   bool IsStopping() const override { return m_stop || ILanguageInvoker::IsStopping(); }
 
-  typedef void (*PythonModuleInitialization)();
-  
+  typedef PyObject* (*PythonModuleInitialization)();
+
 protected:
   // implementation of ILanguageInvoker
   bool execute(const std::string &script, const std::vector<std::string> &arguments) override;
@@ -47,8 +48,8 @@ protected:
   void onExecutionFailed() override;
 
   // custom virtual methods
-  virtual std::map<std::string, PythonModuleInitialization> getModules() const;
-  virtual const char* getInitializationScript() const;
+  virtual std::map<std::string, PythonModuleInitialization> getModules() const = 0;
+  virtual const char* getInitializationScript() const = 0;
   virtual void onInitialization();
   // actually a PyObject* but don't wanna draw Python.h include into the header
   virtual void onPythonModuleInitialization(void* moduleDict);
@@ -67,6 +68,8 @@ private:
   void addPath(const std::string& path); // add path in UTF-8 encoding
   void addNativePath(const std::string& path); // add path in system/Python encoding
   void getAddonModuleDeps(const ADDON::AddonPtr& addon, std::set<std::string>& paths);
+  bool execute(const std::string &script, const std::vector<std::wstring> &arguments);
+  FILE* PyFile_AsFileWithMode(PyObject *py_file, const char *mode);
 
   std::string m_pythonPath;
   void *m_threadState;
