+++ xbmc/interfaces/python/typemaps/python.string.outtm
@@ -18,6 +18,6 @@
  *  <http://www.gnu.org/licenses/>.
  *
  */
-%>${result} = <% 
+%>${result} = <%
   if(method.@feature_python_coerceToUnicode) { %>PyUnicode_DecodeUTF8(${api}.c_str(),${api}.size(),"replace");<% }
-  else { %>PyString_FromStringAndSize(${api}.c_str(), ${api}.length());<% } %>
+  else { %>PyUnicode_FromStringAndSize(${api}.c_str(), ${api}.length());<% } %>
