--- xbmc/interfaces/python3/test/CMakeLists.txt
@@ -1,5 +1,0 @@
-if(PYTHON_FOUND)
-  set(SOURCES TestSwig.cpp)
-
-  core_add_test_library(python_test)
-endif()
