--- xbmc/interfaces/python3/typemaps/python.Alternative.intm
@@ -1,55 +1,0 @@
-<%
-/*
- *      Copyright (C) 2005-2013 Team XBMC
- *      http://xbmc.org
- *
- *  This Program is free software; you can redistribute it and/or modify
- *  it under the terms of the GNU General Public License as published by
- *  the Free Software Foundation; either version 2, or (at your option)
- *  any later version.
- *
- *  This Program is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- *  GNU General Public License for more details.
- *
- *  You should have received a copy of the GNU General Public License
- *  along with XBMC; see the file COPYING.  If not, see
- *  <http://www.gnu.org/licenses/>.
- *
- */
-    boolean ispointer = swigTypeParser.SwigType_ispointer(ltype)
-    String accessor = ispointer ? '->' : '.'
-    int seq = sequence.increment()
-    altAccess = [ 'former', 'later' ]
-    altSwitch = [ 'first', 'second' ]
-
-    List types = swigTypeParser.SwigType_templateparmlist(ltype)
-%>
-    {
-      // we need to check the parameter type and see if it matches
-      PyObject *pyentry_${seq} = ${slarg};
-      try
-      {
-        ${swigTypeParser.SwigType_str(swigTypeParser.SwigType_ltype(types[0]))} entry0_${seq};
-        ${helper.getInConversion(types[0], 'entry0' + '_' + seq, 'pyentry' + '_' + seq, method,
-                                 [ 'sequence' : sequence ])}
-        ${api}${accessor}${altAccess[0]}() = entry0_${seq};
-      }
-      catch (const XBMCAddon::WrongTypeException&)
-      {
-        try
-        {
-          ${swigTypeParser.SwigType_str(swigTypeParser.SwigType_ltype(types[1]))} entry1_${seq};
-          ${helper.getInConversion(types[1], 'entry1' + '_' + seq, 'pyentry' + '_' + seq, method,
-                                   [ 'sequence' : sequence ])}
-          ${api}${accessor}${altAccess[1]}() = entry1_${seq};
-        }
-        catch (const XBMCAddon::WrongTypeException&)
-        {
-          throw XBMCAddon::WrongTypeException("Failed to convert to input type to either a " 
-                                              "${swigTypeParser.SwigType_ltype(types[0])} or a "
-                                              "${swigTypeParser.SwigType_ltype(types[1])}" );
-        }
-      }
-    }
