--- xbmc/interfaces/python3/typemaps/python.Alternative.outtm
@@ -1,46 +1,0 @@
-<%
-/*
- *      Copyright (C) 2005-2013 Team XBMC
- *      http://xbmc.org
- *
- *  This Program is free software; you can redistribute it and/or modify
- *  it under the terms of the GNU General Public License as published by
- *  the Free Software Foundation; either version 2, or (at your option)
- *  any later version.
- *
- *  This Program is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- *  GNU General Public License for more details.
- *
- *  You should have received a copy of the GNU General Public License
- *  along with XBMC; see the file COPYING.  If not, see
- *  <http://www.gnu.org/licenses/>.
- *
- */
-    List types = swigTypeParser.SwigType_templateparmlist(type)
-    boolean ispointer = swigTypeParser.SwigType_ispointer(type)
-    int seq = sequence.increment()
-    String accessor = ispointer ? '->' : '.'
-    altAccess = [ 'former', 'later' ]
-    altSwitch = [ 'first', 'second' ]
-%>
-    WhichAlternative pos = ${api}${accessor}which();
-
-    if (<%if (ispointer) { %>${api} != NULL && <%}%>pos != XBMCAddon::none)
-    { <%
-      types.eachWithIndex { curType, entryIndex -> 
-%>
-      if (pos == XBMCAddon::${altSwitch[entryIndex]})
-      {
-        ${swigTypeParser.SwigType_str(swigTypeParser.SwigType_ltype(curType))}& entry${seq} = ${api}${accessor}${altAccess[entryIndex]}();
-        {
-          ${helper.getOutConversion(curType,result,method,[ 'api' : 'entry' + seq, 'sequence' : sequence ])}
-        }
-      }
-<%
-        }
-%>
-    }
-    else
-      ${result} = Py_None;
