--- xbmc/interfaces/python3/typemaps/python.Tuple.outtm
@@ -1,51 +1,0 @@
-<%
-/*
- *      Copyright (C) 2005-2013 Team XBMC
- *      http://xbmc.org
- *
- *  This Program is free software; you can redistribute it and/or modify
- *  it under the terms of the GNU General Public License as published by
- *  the Free Software Foundation; either version 2, or (at your option)
- *  any later version.
- *
- *  This Program is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- *  GNU General Public License for more details.
- *
- *  You should have received a copy of the GNU General Public License
- *  along with XBMC; see the file COPYING.  If not, see
- *  <http://www.gnu.org/licenses/>.
- *
- */
-    List types = swigTypeParser.SwigType_templateparmlist(type)
-    boolean ispointer = swigTypeParser.SwigType_ispointer(type)
-    int seq = sequence.increment()
-    String accessor = ispointer ? '->' : '.'
-    tupleAccess = [ 'first', 'second', 'third', 'fourth' ]
-%>
-    int vecSize = ${api}${accessor}GetNumValuesSet();
-    ${result} = PyTuple_New(vecSize);
-<%
-    if (ispointer)
-    {
-%>
-    if (${api} != NULL)
-<%  }  // this ends the if (ispointer)
-%>    {
-      PyObject* pyentry${seq}; <%
-      types.eachWithIndex { curType, entryIndex -> 
-%>
-
-      if (vecSize > ${entryIndex})
-      {
-        ${swigTypeParser.SwigType_str(swigTypeParser.SwigType_ltype(curType))}& entry${seq} = ${api}${accessor}${tupleAccess[entryIndex]}();
-        {
-          ${helper.getOutConversion(curType,'result',method,[ 'result' : 'pyentry' + seq, 'api' : 'entry' + seq, 'sequence' : sequence ])}
-        }
-        PyTuple_SetItem(${result}, ${entryIndex}, pyentry${seq});
-      }
-<%
-        }
-%>
-    }
