--- xbmc/interfaces/python3/typemaps/python.map.intm
@@ -1,36 +1,0 @@
-<%
-/*
- *      Copyright (C) 2005-2013 Team XBMC
- *      http://xbmc.org
- *
- *  This Program is free software; you can redistribute it and/or modify
- *  it under the terms of the GNU General Public License as published by
- *  the Free Software Foundation; either version 2, or (at your option)
- *  any later version.
- *
- *  This Program is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- *  GNU General Public License for more details.
- *
- *  You should have received a copy of the GNU General Public License
- *  along with XBMC; see the file COPYING.  If not, see
- *  <http://www.gnu.org/licenses/>.
- *
- */
-    List templateArgs = swigTypeParser.SwigType_templateparmlist(ltype)
-    keytype = templateArgs[0]
-    valtype = templateArgs[1]
-%>
-    {
-      PyObject *pykey, *pyvalue;
-      Py_ssize_t pos = 0;
-      while(PyDict_Next(${slarg}, &pos, &pykey, &pyvalue))
-      {
-        ${swigTypeParser.SwigType_str(swigTypeParser.SwigType_ltype(keytype))} key;
-        ${swigTypeParser.SwigType_str(swigTypeParser.SwigType_ltype(valtype))} value;
-        ${helper.getInConversion(keytype, 'key', 'pykey', method)}
-        ${helper.getInConversion(valtype, 'value', 'pyvalue' ,method)}
-        ${api}[key] = value;
-      }
-    }
