--- xbmc/interfaces/python3/typemaps/python.vector.intm
@@ -1,47 +1,0 @@
-<%
-/*
- *      Copyright (C) 2005-2013 Team XBMC
- *      http://xbmc.org
- *
- *  This Program is free software; you can redistribute it and/or modify
- *  it under the terms of the GNU General Public License as published by
- *  the Free Software Foundation; either version 2, or (at your option)
- *  any later version.
- *
- *  This Program is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- *  GNU General Public License for more details.
- *
- *  You should have received a copy of the GNU General Public License
- *  along with XBMC; see the file COPYING.  If not, see
- *  <http://www.gnu.org/licenses/>.
- *
- */
-    List templateArgs = swigTypeParser.SwigType_templateparmlist(ltype)
-    vectype = templateArgs[0]
-    boolean ispointer = swigTypeParser.SwigType_ispointer(type)
-    String accessor = ispointer ? '->' : '.'
-    int seq = sequence.increment()
-%>
-    if (${slarg})
-    {
-      bool isTuple = PyObject_TypeCheck(${slarg},&PyTuple_Type);
-      if (!isTuple && !PyObject_TypeCheck(${slarg},&PyList_Type))
-        throw WrongTypeException("The parameter \"${api}\" must be either a Tuple or a List.");
-
-      <%  if (ispointer) print("${api} = new std::vector<${swigTypeParser.SwigType_str(vectype)}>();") %>
-      PyObject *pyentry${seq} = NULL;
-      int vecSize = (isTuple ? PyTuple_Size(${slarg}) : PyList_Size(${slarg}));
-      for(int i = 0; i < vecSize; i++)
-      {
-        pyentry${seq} = (isTuple ? PyTuple_GetItem(${slarg}, i) : PyList_GetItem(${slarg}, i));
-        ${swigTypeParser.SwigType_str(swigTypeParser.SwigType_ltype(vectype))} entry${seq};
-        ${helper.getInConversion(vectype, 'entry' + seq, 'pyentry' + seq, method,
-                                 [ 'type' : vectype,
-                                   'ltype' : swigTypeParser.SwigType_ltype(vectype),
-                                   'sequence' : sequence
-                                   ])}
-        ${api}${accessor}push_back(entry${seq});
-      }
-    }
